+++ src/write.c
@@ -425,7 +425,7 @@ static avifResult avifItemPropertyDedupFinish(avifItemPropertyDedup * dedup, avi
             !memcmp(&outputStream->raw->data[property->offset], dedup->buffer.data, newPropertySize)) {
             // We've already written this exact property, reuse it
             propertyIndex = property->index;
-            assert(propertyIndex != 0);
+            AVIF_ASSERT_OR_RETURN(propertyIndex != 0);
             break;
         }
     }
@@ -639,7 +639,7 @@ static avifResult avifEncoderWriteColorProperties(avifRWStream * outputStream,
     // of an already stored property.
     avifRWStream * dedupStream = outputStream;
     if (dedup) {
-        assert(ipma);
+        AVIF_ASSERT_OR_RETURN(ipma);
 
         // Use the dedup's temporary stream for box writes.
         dedupStream = &dedup->s;
@@ -821,7 +821,7 @@ static avifResult avifEncoderWriteTrackMetaBox(avifEncoder * encoder, avifRWStre
             continue;
         }
 
-        assert(!item->hiddenImage);
+        AVIF_ASSERT_OR_RETURN(!item->hiddenImage);
         avifBoxMarker infe;
         AVIF_CHECKRES(avifRWStreamWriteFullBox(s, "infe", AVIF_BOX_SIZE_TBD, 2, 0, &infe));
         AVIF_CHECKRES(avifRWStreamWriteU16(s, item->id));                             // unsigned int(16) item_ID;
@@ -1022,36 +1022,23 @@ static avifResult avifEncoderDataCreateXMPItem(avifEncoderData * data, const avi
 }
 
 // Same as avifImageCopy() but pads the dstImage with border pixel values to reach dstWidth and dstHeight.
-// Returns NULL if a memory allocation failed.
-static avifImage * avifImageCopyAndPad(const avifImage * srcImage, uint32_t dstWidth, uint32_t dstHeight)
+static avifResult avifImageCopyAndPad(avifImage * const dstImage, const avifImage * srcImage, uint32_t dstWidth, uint32_t dstHeight)
 {
-    avifImage * dstImage = avifImageCreateEmpty();
-    if (!dstImage) {
-        return NULL;
-    }
-    // Copy all fields but do not allocate.
-    if (avifImageCopy(dstImage, srcImage, (avifPlanesFlag)0) != AVIF_RESULT_OK) {
-        avifImageDestroy(dstImage);
-        return NULL;
-    }
-    assert(dstWidth >= srcImage->width);
-    assert(dstHeight >= srcImage->height);
+    AVIF_ASSERT_OR_RETURN(dstImage);
+    AVIF_ASSERT_OR_RETURN(!dstImage->width && !dstImage->height); // dstImage is not set yet.
+    AVIF_ASSERT_OR_RETURN(dstWidth >= srcImage->width);
+    AVIF_ASSERT_OR_RETURN(dstHeight >= srcImage->height);
+
+    // Copy all fields but do not allocate the planes.
+    AVIF_CHECKRES(avifImageCopy(dstImage, srcImage, (avifPlanesFlag)0));
     dstImage->width = dstWidth;
     dstImage->height = dstHeight;
 
     if (srcImage->yuvPlanes[AVIF_CHAN_Y]) {
-        const avifResult allocationResult = avifImageAllocatePlanes(dstImage, AVIF_PLANES_YUV);
-        if (allocationResult != AVIF_RESULT_OK) {
-            avifImageDestroy(dstImage);
-            return NULL;
-        }
+        AVIF_CHECKRES(avifImageAllocatePlanes(dstImage, AVIF_PLANES_YUV));
     }
     if (srcImage->alphaPlane) {
-        const avifResult allocationResult = avifImageAllocatePlanes(dstImage, AVIF_PLANES_A);
-        if (allocationResult != AVIF_RESULT_OK) {
-            avifImageDestroy(dstImage);
-            return NULL;
-        }
+        AVIF_CHECKRES(avifImageAllocatePlanes(dstImage, AVIF_PLANES_A));
     }
     const avifBool usesU16 = avifImageUsesU16(srcImage);
     for (int plane = AVIF_CHAN_Y; plane <= AVIF_CHAN_A; ++plane) {
@@ -1091,7 +1078,7 @@ static avifImage * avifImageCopyAndPad(const avifImage * srcImage, uint32_t dstW
             dstRow += dstRowBytes;
         }
     }
-    return dstImage;
+    return AVIF_RESULT_OK;
 }
 
 static int avifQualityToQuantizer(int quality, int minQuantizer, int maxQuantizer)
@@ -1242,9 +1229,9 @@ static avifResult avifValidateGrid(uint32_t gridCols,
     const avifImage * bottomRightCell = cellImages[cellCount - 1];
 #if defined(AVIF_ENABLE_EXPERIMENTAL_GAIN_MAP)
     if (validateGainMap) {
-        assert(firstCell->gainMap && firstCell->gainMap->image);
+        AVIF_ASSERT_OR_RETURN(firstCell->gainMap && firstCell->gainMap->image);
         firstCell = firstCell->gainMap->image;
-        assert(bottomRightCell->gainMap && bottomRightCell->gainMap->image);
+        AVIF_ASSERT_OR_RETURN(bottomRightCell->gainMap && bottomRightCell->gainMap->image);
         bottomRightCell = bottomRightCell->gainMap->image;
     }
 #endif
@@ -1256,7 +1243,7 @@ static avifResult avifValidateGrid(uint32_t gridCols,
         const avifImage * cellImage = cellImages[cellIndex];
 #if defined(AVIF_ENABLE_EXPERIMENTAL_GAIN_MAP)
         if (validateGainMap) {
-            assert(cellImage->gainMap && cellImage->gainMap->image);
+            AVIF_ASSERT_OR_RETURN(cellImage->gainMap && cellImage->gainMap->image);
             cellImage = cellImage->gainMap->image;
         }
 #endif
@@ -1530,12 +1517,12 @@ static avifResult avifEncoderAddImageInternal(avifEncoder * encoder,
             uint16_t alphaItemID;
             AVIF_CHECKRES(avifEncoderAddImageItems(encoder, gridCols, gridRows, gridWidth, gridHeight, AVIF_ITEM_ALPHA, &alphaItemID));
             avifEncoderItem * alphaItem = avifEncoderDataFindItemByID(encoder->data, alphaItemID);
-            assert(alphaItem);
+            AVIF_ASSERT_OR_RETURN(alphaItem);
             alphaItem->irefType = "auxl";
             alphaItem->irefToID = colorItemID;
             if (encoder->data->imageMetadata->alphaPremultiplied) {
                 avifEncoderItem * colorItem = avifEncoderDataFindItemByID(encoder->data, colorItemID);
-                assert(colorItem);
+                AVIF_ASSERT_OR_RETURN(colorItem);
                 colorItem->irefType = "prem";
                 colorItem->irefToID = alphaItemID;
             }
@@ -1573,17 +1560,17 @@ static avifResult avifEncoderAddImageInternal(avifEncoder * encoder,
             AVIF_CHECKRES(
                 avifEncoderAddImageItems(encoder, gridCols, gridRows, gainMapGridWidth, gainMapGridHeight, AVIF_ITEM_GAIN_MAP, &gainMapItemID));
             avifEncoderItem * gainMapItem = avifEncoderDataFindItemByID(encoder->data, gainMapItemID);
-            assert(gainMapItem);
+            AVIF_ASSERT_OR_RETURN(gainMapItem);
             gainMapItem->hiddenImage = AVIF_TRUE;
 
             // Set the color item and gain map item's dimgFromID value to point to the tone mapped item.
             // The color item shall be first, and the gain map second. avifEncoderFinish() writes the
             // dimg item references in item id order, so as long as colorItemID < gainMapItemID, the order
             // will be correct.
-            assert(colorItemID < gainMapItemID);
+            AVIF_ASSERT_OR_RETURN(colorItemID < gainMapItemID);
             avifEncoderItem * colorItem = avifEncoderDataFindItemByID(encoder->data, colorItemID);
-            assert(colorItem);
-            assert(colorItem->dimgFromID == 0); // Our internal API only allows one dimg value per item.
+            AVIF_ASSERT_OR_RETURN(colorItem);
+            AVIF_ASSERT_OR_RETURN(colorItem->dimgFromID == 0); // Our internal API only allows one dimg value per item.
             colorItem->dimgFromID = toneMappedItemID;
             gainMapItem->dimgFromID = toneMappedItemID;
         }
@@ -1643,11 +1630,9 @@ static avifResult avifEncoderAddImageInternal(avifEncoder * encoder,
         //     }
         // Check whether it is safe to cast width and height to uint16_t. The maximum width and
         // height of an AV1 frame are 65536, which just exceeds uint16_t.
-        assert(encoder->data->items.count > 0);
+        AVIF_ASSERT_OR_RETURN(encoder->data->items.count > 0);
         const avifImage * imageMetadata = encoder->data->imageMetadata;
-        if ((imageMetadata->width > 65535) || (imageMetadata->height > 65535)) {
-            return AVIF_RESULT_INVALID_ARGUMENT;
-        }
+        AVIF_CHECKERR(imageMetadata->width <= 65535 && imageMetadata->height <= 65535, AVIF_RESULT_INVALID_ARGUMENT);
     }
 
     // -----------------------------------------------------------------------
@@ -1660,17 +1645,20 @@ static avifResult avifEncoderAddImageInternal(avifEncoder * encoder,
             const avifImage * firstCellImage = firstCell;
 #if defined(AVIF_ENABLE_EXPERIMENTAL_GAIN_MAP)
             if (item->itemCategory == AVIF_ITEM_GAIN_MAP) {
-                assert(cellImage->gainMap && cellImage->gainMap->image);
+                AVIF_ASSERT_OR_RETURN(cellImage->gainMap && cellImage->gainMap->image);
                 cellImage = cellImage->gainMap->image;
-                assert(firstCell->gainMap && firstCell->gainMap->image);
+                AVIF_ASSERT_OR_RETURN(firstCell->gainMap && firstCell->gainMap->image);
                 firstCellImage = firstCell->gainMap->image;
             }
 #endif
             avifImage * paddedCellImage = NULL;
             if ((cellImage->width != firstCellImage->width) || (cellImage->height != firstCellImage->height)) {
-                paddedCellImage = avifImageCopyAndPad(cellImage, firstCellImage->width, firstCellImage->height);
-                if (!paddedCellImage) {
-                    return AVIF_RESULT_OUT_OF_MEMORY;
+                paddedCellImage = avifImageCreateEmpty();
+                AVIF_CHECKERR(paddedCellImage, AVIF_RESULT_OUT_OF_MEMORY);
+                const avifResult result = avifImageCopyAndPad(paddedCellImage, cellImage, firstCellImage->width, firstCellImage->height);
+                if (result != AVIF_RESULT_OK) {
+                    avifImageDestroy(paddedCellImage);
+                    return result;
                 }
                 cellImage = paddedCellImage;
             }
@@ -1804,7 +1792,7 @@ static avifResult avifEncoderWriteMediaDataBox(avifEncoder * encoder,
             if ((encoder->extraLayerCount > 0) && (item->encodeOutput->samples.count > 0)) {
                 // Interleave - Pick out AV1 items and interleave them later.
                 // We always interleave all AV1 items for layered images.
-                assert(item->encodeOutput->samples.count == item->mdatFixups.count);
+                AVIF_ASSERT_OR_RETURN(item->encodeOutput->samples.count == item->mdatFixups.count);
 
                 avifEncoderItemReference * ref = (item->itemCategory == AVIF_ITEM_ALPHA) ? avifArrayPush(layeredAlphaItems)
                                                                                          : avifArrayPush(layeredColorItems);
@@ -1905,7 +1893,7 @@ static avifResult avifEncoderWriteMediaDataBox(avifEncoder * encoder,
             ++layerIndex;
         } while (hasMoreSample);
 
-        assert(layerIndex <= AVIF_MAX_AV1_LAYER_COUNT);
+        AVIF_ASSERT_OR_RETURN(layerIndex <= AVIF_MAX_AV1_LAYER_COUNT);
     }
     avifRWStreamFinishBox(s, mdat);
     return AVIF_RESULT_OK;
@@ -1964,7 +1952,7 @@ static avifResult avifImageWriteExtendedMeta(const avifImage * imageMetadata, av
                                                    ((imageMetadata->transformFlags & AVIF_TRANSFORM_IROT) ? 1 : 0) +
                                                    ((imageMetadata->transformFlags & AVIF_TRANSFORM_IMIR) ? 1 : 0);
             const uint8_t ipmaCount = numNonEssentialProperties + numEssentialProperties;
-            assert(ipmaCount >= 1);
+            AVIF_ASSERT_OR_RETURN(ipmaCount >= 1);
 
             // Only add properties to the primary item.
             AVIF_CHECKRES(avifRWStreamWriteU32(stream, 1)); // unsigned int(32) entry_count;
@@ -2074,15 +2062,15 @@ static avifResult avifEncoderWriteCondensedImageBox(avifEncoder * encoder, avifR
         avifEncoderItem * item = &encoder->data->items.item[itemIndex];
         if (item->id == encoder->data->primaryItemID) {
             colorItem = item;
-            assert(colorItem->encodeOutput->samples.count == 1);
+            AVIF_ASSERT_OR_RETURN(colorItem->encodeOutput->samples.count == 1);
         } else if (item->itemCategory == AVIF_ITEM_ALPHA && item->irefToID == encoder->data->primaryItemID) {
-            assert(!alphaItem);
+            AVIF_ASSERT_OR_RETURN(!alphaItem);
             alphaItem = item;
-            assert(alphaItem->encodeOutput->samples.count == 1);
+            AVIF_ASSERT_OR_RETURN(alphaItem->encodeOutput->samples.count == 1);
         }
     }
 
-    assert(colorItem);
+    AVIF_ASSERT_OR_RETURN(colorItem);
     const avifRWData * colorData = &colorItem->encodeOutput->samples.sample[0].data;
     const avifRWData * alphaData = alphaItem ? &alphaItem->encodeOutput->samples.sample[0].data : NULL;
 
@@ -2146,7 +2134,7 @@ static avifResult avifEncoderWriteCondensedImageBox(avifEncoder * encoder, avifR
                                                    // unsigned int(32) infe_type;
                                                    // unsigned int(32) codec_config_type;
     AVIF_CHECKRES(avifRWStreamWriteVarInt(s, 4));  // varint main_item_codec_config_size;
-    assert(colorData->size >= 1);
+    AVIF_ASSERT_OR_RETURN(colorData->size >= 1);
     AVIF_CHECKRES(avifRWStreamWriteVarInt(s, (uint32_t)colorData->size - 1)); // varint main_item_data_size_minus_one;
 
     AVIF_CHECKRES(avifRWStreamWriteBits(s, alphaItem ? 1 : 0, 1)); // unsigned int(1) has_alpha;
@@ -2253,7 +2241,7 @@ static avifResult avifRWStreamWriteProperties(avifItemPropertyDedup * const dedu
         if (isToneMappedImage) {
             itemMetadata = altImageMetadata;
         } else if (item->itemCategory == AVIF_ITEM_GAIN_MAP) {
-            assert(itemMetadata->gainMap && itemMetadata->gainMap->image);
+            AVIF_ASSERT_OR_RETURN(itemMetadata->gainMap && itemMetadata->gainMap->image);
             itemMetadata = itemMetadata->gainMap->image;
         }
 #else
@@ -2761,7 +2749,7 @@ avifResult avifEncoderFinish(avifEncoder * encoder, avifRWData * output)
             durationInTimescales = AVIF_INDEFINITE_DURATION64;
         } else {
             uint64_t loopCount = encoder->repetitionCount + 1;
-            assert(framesDurationInTimescales != 0);
+            AVIF_ASSERT_OR_RETURN(framesDurationInTimescales != 0);
             if (loopCount > UINT64_MAX / framesDurationInTimescales) {
                 // The multiplication will overflow uint64_t.
                 return AVIF_RESULT_INVALID_ARGUMENT;